Move modules into specific packages based on category.
